Move new Stream to new folder
